Added images to README and changed 'offline_stuff' folder to 'lab'
