--- offline_stuff/README.txt
@@ -1,2 +1,0 @@
-## Folder for storing preparatory material generated by Hadrien.
-## Enjoy!
